+++ piezas/horse.py
@@ -1,8 +1,9 @@
 
 from piezas.piece import Piece
 class Horse(Piece):
-    black_str ="♘" 
-    white_str ="♞"  
+    black_str ="♞"  
+    white_str ="♘" 
+    
     def __init__(self, color):
         super().__init__(color)